refactor: annotate cycle analyzer
